[add] 컴포넌트 분리 (UserFeeds, UserInfo 컴포넌트 생성) #91
